Inlined "shared" behavior which was only being used by a single benchmark suite.
